requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
airtable-python-wrapper==0.15.3
async-timeout==4.0.2
attrs==22.2.0
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==2.1.1
discord==2.1.0
discord.py==2.1.0
et-xmlfile==1.1.0
Faker==16.6.1
frozenlist==1.3.3
google-api-core==2.11.0
google-api-python-client==2.74.0
google-auth==2.16.0
google-auth-httplib2==0.1.0
googleapis-common-protos==1.58.0
httplib2==0.21.0
idna==3.4
multidict==6.0.4
numpy==1.24.1
openpyxl==3.1.0
pandas==1.5.3
protobuf==4.21.12
pyairtable==1.4.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.21.1
pytz==2022.7.1
requests==2.28.2
rsa==4.9
six==1.16.0
uritemplate==4.1.1
urllib3==1.26.14
yarl==1.8.2
